+++ drivers/qmimodem/qmi.h
@@ -97,8 +97,8 @@ bool qmi_device_discover(struct qmi_device *device, qmi_discover_func_t func,
 bool qmi_device_shutdown(struct qmi_device *device, qmi_shutdown_func_t func,
 				void *user_data, qmi_destroy_func_t destroy);
 
-bool qmi_device_has_service(struct qmi_device *device, uint8_t type);
-bool qmi_device_get_service_version(struct qmi_device *device, uint8_t type,
+bool qmi_device_has_service(struct qmi_device *device, uint16_t type);
+bool qmi_device_get_service_version(struct qmi_device *device, uint16_t type,
 					uint16_t *major, uint16_t *minor);
 
 bool qmi_device_sync(struct qmi_device *device,
@@ -158,10 +158,10 @@ typedef void (*qmi_result_func_t)(struct qmi_result *result, void *user_data);
 typedef void (*qmi_create_func_t)(struct qmi_service *service, void *user_data);
 
 bool qmi_service_create(struct qmi_device *device,
-				uint8_t type, qmi_create_func_t func,
+				uint16_t type, qmi_create_func_t func,
 				void *user_data, qmi_destroy_func_t destroy);
 bool qmi_service_create_shared(struct qmi_device *device,
-				uint8_t type, qmi_create_func_t func,
+				uint16_t type, qmi_create_func_t func,
 				void *user_data, qmi_destroy_func_t destroy);
 
 struct qmi_service *qmi_service_ref(struct qmi_service *service);
